fix: move all lib or util into util folder
