+++ src/components/ui/dropdown-menu.tsx
@@ -2,7 +2,7 @@ import * as DropdownMenuPrimitive from "@radix-ui/react-dropdown-menu";
 import { Check, ChevronRight, Circle } from "lucide-react";
 import * as React from "react";
 
-import { cn } from "@/lib/utils";
+import { cn } from "@/utils/classname";
 
 const DropdownMenu = DropdownMenuPrimitive.Root;
