+++ src/components/ui/hover-card.tsx
@@ -1,7 +1,7 @@
 import * as HoverCardPrimitive from "@radix-ui/react-hover-card";
 import * as React from "react";
 
-import { cn } from "@/lib/utils";
+import { cn } from "@/utils/classname";
 
 const HoverCard = HoverCardPrimitive.Root;
