+++ src/components/ui/tabs.tsx
@@ -1,7 +1,7 @@
 import * as TabsPrimitive from "@radix-ui/react-tabs";
 import * as React from "react";
 
-import { cn } from "@/lib/utils";
+import { cn } from "@/utils/classname";
 
 const Tabs = TabsPrimitive.Root;
