+++ src/utils/animation.ts
